==> src/1-agents-process.ipynb <==
# Agents processing
1) Generate a subset of agents who live within Västra Gätalands Region.

%load_ext autoreload
%autoreload 2
%cd D:\synthetic-sweden

# Load libs
import pandas as pd
import geopandas as gpd
from tqdm import tqdm
import dataworkers as dw

## Load boundary

boundary = gpd.read_file('dbs/municipalities/sweden_komk.shp')
mun_in_boundary = list(boundary.loc[boundary['county'] == 'Västra Götaland', 'KnKod'].values)

## Load agents and their daily plans

df_pop = pd.read_pickle('dbs/agents/syn_pop_all.pkl')
df_pop.head(3)

df_pop_sub = df_pop.loc[df_pop['kommun'].isin(mun_in_boundary), :]
print("Agents live in Västra Götalands: %s"%len(df_pop_sub))

df_plan = pd.read_pickle('dbs/agents/df_act_plan.pkl')
df_plan.head(3)

df_plan_sub = df_plan.loc[df_plan['PId'].isin(df_pop_sub.PId), :]
print("Agents live in Västra Götalands have activities: %s"%len(df_plan_sub))

## Keep agents and their activities if they all locate within the boundary
This is not necessary anymore for the current version of case study.

gdf = dw.df2gdf_point(df_plan_sub, 'POINT_X', 'POINT_Y', crs=4326, drop=False)
gdf.loc[:, 'boundary'] = gdf.apply(lambda row: boundary.contains(row.geometry), axis=1)
gdf.head(2)

tqdm.pandas()
df_id2select = gdf.groupby('PId').progress_apply(lambda g: all(g['boundary'])).reset_index(name='select')
id2select = df_id2select.loc[df_id2select['select'], 'PId'].values

## Save agents and their plans for a subset in VRG

df_plan_sub.to_pickle("dbs/agents/df_act_plan_vgr.pkl")
df_pop_sub.to_pickle("dbs/agents/syn_pop_vgr.pkl")
