==> src/2-matsim-data-preparation.ipynb <==
# MATSim input data preparation
Prepare road network and synthetic population.

%load_ext autoreload
%autoreload 2
%cd D:\synthetic-sweden
%matplotlib inline

# Load libs
import pandas as pd
import numpy as np
from tqdm import tqdm
import dataworkers as dw
import matsim
import gzip

## 1. Network preparation

### Load boundary
Create boundary using QGIS focusing on VGR.

### Prepare command line
#### .osm.pbf -> .xml.gz

file_java = 'D:\\matsim-code-examples-13.x\\src\\main\\java\\org\\matsim\\codeexamples\\network'
print(file_java)

Epsg = "EPSG:3006"
input = "D:\\synthetic-sweden\\dbs\\network\\sweden-latest.osm.pbf"
filter = "D:\\synthetic-sweden\\dbs\\network\\vgr_filter.shp" # Created in QGIS
output = "D:\\synthetic-sweden\\dbs\\network\\vgr.xml.gz"

### Convert boundary into poly for osmosis filtering

dw.shp2poly(filter, targetfile="D:\\synthetic-sweden\\dbs\\network\\vgr_filter.poly")

## 2. Population preparation

# h-Home, w-Work, o-Other, s-School
activity_purpose_dict = {1: 'h', 4: 'w', 10: 'o', 6: 's'}

### Load agents who live in VGR

df_pop = pd.read_pickle('dbs/agents/syn_pop_vgr.pkl')
df_pop.head(3)

df_plan = pd.read_pickle('dbs/agents/df_act_plan_vgr.pkl')
df_plan['act_purpose'] = df_plan['act_purpose'].map(activity_purpose_dict)
df_plan.head(3)

print(df_plan.act_purpose.unique())

print(df_plan['mode'].unique())

### Sample agents for simulation
Create a subset of car users from all VG population.

agents_car_users = df_plan.loc[df_plan['mode'].isin(['Car', 'CarPassenger']), 'PId'].unique()
df_plan_sim = df_plan.loc[df_plan['PId'].isin(agents_car_users), :]

#### Dump the agents' activity plans into MATSim format
Excluding those who

1) do not move (stay home),
2) start and end locations are not the same

purpose_dict = {'h': 'home', 'o': 'other', 'w': 'work', 's': 'school'}
mode_dict = {'Car': 'car', 'CarPassenger': 'car', 'Bike': 'bike',
             'Walking': 'walk', 'PublicTransport': 'pt'}

def data2xml(data):
    num_activities = len(data)
    # Modes between activities
    mode_array = data['mode'].values[1:]
    act_purpose_array = data['act_purpose'].values
    X_array = data['POINT_X_sweref99'].values
    Y_array = data['POINT_Y_sweref99'].values
    act_end_array = data['act_end'].values
    # Process the case where the last activity crosses 0
    if act_end_array[-2] < 3:
        act_purpose_array = np.concatenate((act_purpose_array[-2], act_purpose_array[:-1]), axis=None)
        X_array = np.concatenate((X_array[-2], X_array[:-1]), axis=None)
        Y_array = np.concatenate((Y_array[-2], Y_array[:-1]), axis=None)
        act_end_array = np.concatenate((act_end_array[-2], act_end_array[:-1]), axis=None)
        act_end_array[-1] = 23.99
    # Excluding those who 1) do not move (stay home),
    # 2) start and end locations are not the same
    if (num_activities > 2) & (act_purpose_array[0] == act_purpose_array[-1]):
        writer.start_person(person_id=data['PId'].values[0])
        writer.start_plan(selected=True)
        for i in range(num_activities):
            # The last activity is always home
            if i == num_activities - 1:
                writer.add_activity(type=purpose_dict[act_purpose_array[i]],
                                    x=X_array[i],
                                    y=Y_array[i])
            else:
                writer.add_activity(type=purpose_dict[act_purpose_array[i]],
                        x=X_array[i],
                        y=Y_array[i],
                        end_time=act_end_array[i]*3600)
                writer.add_leg(mode=mode_dict[mode_array[i]])
        writer.end_plan()
        writer.end_person()

with gzip.open("dbs/agents/plans_cars.xml.gz", 'wb+') as f_write:
    writer = matsim.writers.PopulationWriter(f_write)
    writer.start_population()
    tqdm.pandas()
    df_plan_sim.groupby('PId').progress_apply(data2xml)
    writer.end_population()
