==> src/12-charge-points-comparison.ipynb <==
# Compare the simulation results with existing charging infrastructure in Sweden

%load_ext autoreload
%autoreload 2
%cd D:\synthetic-sweden
%matplotlib inline

import pandas as pd
import geopandas as gpd
import dataworkers as dw
scenario = 'scenario_vg_car'

## Load today's infrastructure

df = pd.read_excel('dbs/ev/charging_sites.xlsx')
df.head()

df = df.loc[:, ['id', 'lat', 'lng', 'pw', 'n']]
df.loc[:, 'fast'] = df.apply(lambda row: row['n'] if row['pw'] > 22 else 0, axis=1)

gdf = dw.df2gdf_point(df, 'lng', 'lat', crs=4326, drop=True).to_crs(3006)
gdf.head()

## Load simulation results

df_sim = pd.read_csv(f"results/{scenario}_stats_charger_5days.csv")
df_sim = df_sim.loc[df_sim['power_fast'] == 50, :]
df_sim = df_sim.groupby(['deso', 'Charging_type'])['number'].sum().reset_index()
df_sim.head()

zones = gpd.read_file('dbs/DeSO/DeSO_2018_v2.shp')
zones = zones.loc[zones['deso'].isin(df_sim['deso'].unique()), :]

## Calculate existing number of charging points

gdf = gpd.sjoin(gdf, zones.loc[:, ['deso', 'befolkning', 'geometry']])
gdf.head()

df_inf = gdf.groupby('deso')['n'].sum().reset_index()
df_inf.head()

df = pd.merge(df_sim, df_inf, on='deso', how='outer')
df = df.fillna(0)
df.columns = ['deso', 'Charging_type', 'sim', 'gt']
df.head()

# Sale up to total VG residents
df.sim = df.sim * (1/0.35) # The results are based on 35% VG car users.
df.to_csv(f'results/{scenario}_inf_comp.csv', index=False)
