==> src/14-sensitivity-results-summary.ipynb <==
# Summary of sensitivity results test

%load_ext autoreload
%autoreload 2
%cd D:\synthetic-sweden
%matplotlib inline

import pandas as pd
from tqdm import tqdm
pd.options.display.float_format = '{:,.2f}'.format
scenario = 'scenario_vg_car'

## 1. Individual statistics

df_indi = pd.read_csv(f'results/sensitivity/{scenario}_stats_indi_agg_5days.csv')
tqdm.pandas()
df_indi.loc[:, 'Charging_type'] = df_indi.progress_apply(lambda row: str(int(row['charging_type'])) + ' (' + str(row['type1']) + ')' if row['charging_type'] in (1, 2) else str(int(row['charging_type'])) + ' (0.9)', axis=1)
df_indi.head()

### 1.1 Share of failed agents

table = pd.pivot_table(df_indi, values='failure_rate', index=['Charging_type', 'power_fast'],
                    columns=['home_charger'])
table.to_clipboard()
table

### 1.2 Share of agents who accessed daytime charging

table = pd.pivot_table(df_indi, values='charging_share', index=['Charging_type', 'power_fast'],
                    columns=['home_charger'])
table.to_clipboard()
table

### 1.3 Charging time and ratio, energy / person and total energy

df_indi_energy = pd.read_csv(f'results/sensitivity/{scenario}_stats_indi_5days.csv')
df_indi_energy.head()

def data_summary(data):
    data2compute_inter = data.loc[data['charging_time_inter']>0, :]
    data2compute_fast = data.loc[data['charging_time_fast']>0, :]
    sum_dict = dict(
        charging_time_inter = data2compute_inter['charging_time_inter'].median(),
        charging_time_ratio_inter = data2compute_inter['charging_time_ratio'].median()*100,
        charging_time_fast = data2compute_fast['charging_time_fast'].median(),
        charging_time_ratio_fast = data2compute_fast['charging_time_ratio'].median()*100,
        energy_daytime=(data['charging_energy_fast'].sum() + data['charging_energy_inter'].sum())/10**3,
        energy_overnight=data['charging_energy_overnight'].sum()/10**3
    )
    return pd.Series(sum_dict)

df_indi_e = df_indi_energy.groupby(['Charging_type', 'power_fast', 'home_charger']).apply(data_summary).reset_index() # .to_clipboard()
table = pd.pivot_table(df_indi_e, values=['charging_time_inter', 'charging_time_ratio_inter', 'charging_time_fast', 'charging_time_ratio_fast',
                                          'energy_daytime', 'energy_overnight'], index=['Charging_type', 'power_fast'],
                    columns=['home_charger'])
table.to_clipboard()
table

## 2. Spatiotemporal statistics

### 2.1 Charger in DeSO zones and aggregate statistics

df_charger = pd.read_csv(f'results/sensitivity/{scenario}_stats_charger_5days.csv')
df_charger.head()

df_agg = df_charger.groupby(['Charging_type', 'Purpose', 'power_fast', 'charger'])['number'].sum().reset_index().drop_duplicates(subset=['Charging_type', 'Purpose', 'power_fast', 'charger'])
table = pd.pivot_table(df_agg, values='number', index=['Charging_type','power_fast', 'Purpose'], columns=['charger'])
table.to_clipboard()
table

Overnight chargers

df_indi_raw = pd.read_csv(f'results/sensitivity/{scenario}_indi_raw_5days.csv')
df_indi_raw.head()

df_indi_raw.groupby(['charging_type', 'paraset'])['home_charger', 'residential_charger'].sum()

### 2.2 Temporal power demand
Visualisation to present the results.

df_power = pd.read_csv(f'results/sensitivity/{scenario}_stats_power_5days.csv')
df_power.head(70)

## 3. Compare the simulation results with existing charging infrastructure in Sweden

df_inf_comp = pd.read_csv(f'results/sensitivity/{scenario}_inf_comp.csv')
df_inf_comp.head()

1) Number of zones need more infrastructure, and
2) Maximum number of additional charging points needed.

import numpy as np
def stats_inf_comp(data):
    zone_share = len(data.loc[data['sim'] > data['gt'], :]) / len(data) * 100
    addi_num_max = max(data['sim'] - data['gt'])
    addi_num_median = np.median(data['sim'] - data['gt'])
    return pd.Series({'zone_share': zone_share, 'addi_num_max': addi_num_max, 'addi_num_median': addi_num_median})
df_inf_comp.groupby(['Charging_type', 'power_fast']).apply(stats_inf_comp).to_clipboard()
df_inf_comp.groupby(['Charging_type', 'power_fast']).apply(stats_inf_comp)
